+++ src/constants/localStorage.js
@@ -1,9 +1,19 @@
-export const TOKEN = 'token'
-export const ACCOUNT_INFO = 'accountInfo'
-export const SETUP_TIME = 'setupTime'
-export const USER_ID = 'userID'
-export const ACCOUNT_ID = 'accountID'
-export const keys = [TOKEN, ACCOUNT_INFO, SETUP_TIME, USER_ID, ACCOUNT_ID]
+// export const TOKEN = 'token'
+// export const ACCOUNT_INFO = 'accountInfo'
+// export const SETUP_TIME = 'setupTime'
+// export const ID = 'userID'
+// export const ACCOUNT_ID = 'accountID'
+// export const keys = [TOKEN, ACCOUNT_INFO, SETUP_TIME, USER_ID, ACCOUNT_ID]
 
+
+export const AUTH_INFO = 'authInfo'
+export const EMAIL = 'email'
+export const FIRST_NAME = 'firstname'
+export const ID = 'id'
+export const TOKEN = 'jwt'
+export const LAST_NAME = 'lastname'
+export const PRIVILEGE = 'privilege'
+export const SET_UP_TIME = 'setupTime'
+export const STATUS = 'status'
 
 // SETUP_TIME is used for responseConverter (auth string to obj) originally