+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ### Unrelease
+- Remove redundance
+- Add bookmarks related actions/reducers from @twreporter/redux
+- Add bookmark functional icons from twreporter-react(main website) and export it as widget. ( The icons are fully functional )
+
 
 ### 2.1.5
 - Rasterize facebook_button_logo.svg file